requirements.txt
certifi==2020.6.20
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.3.1
mysql-connector-python==8.0.21
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
protobuf==3.13.0
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
python-dateutil==2.8.1
six==1.15.0
